Enable Swagger , Enable Docker app properties
