+++ docker-compose.yml
@@ -1,8 +1,48 @@
-version : '3.6'
-
+version: '3'
 services:
-  my-docker-compose-app:
-    image: my-docker-compose-app
-    build: .
+  vms-compose:
+    image: vms-app-image-1
+    container_name: vms_container
+    build: 
+      context: .
     ports:
-      - 9090:8081
+      - 8080:8080
+#    networks:
+#      - vms-mysql-net
+    depends_on:
+      - mysql-docker-container
+#    restart: always
+ 
+ 
+  mysql-docker-container:
+    image: mysql:5.7
+    environment:
+      - MYSQL_ROOT_PASSWORD=adminpassword
+      - MYSQL_DATABASE=test
+      - MYSQL_USER=testuser
+      - MYSQL_PASSWORD=testpassword
+    ports:
+      # <Port exposed> : < MySQL Port running inside container>
+      - 3307:3306
+#    networks:
+#      - vms-mysql-net
+ 
+#  db:
+#    image: mysql:5.7
+#    container_name: DB_container
+##    restart: always
+#    environment:
+#      MYSQL_DATABASE: 'db'
+#      # So you don't have to use root, but you can if you like
+#      MYSQL_USER: user
+#      # You can use whatever password you like
+#      MYSQL_PASSWORD: password
+#      # Password for root access
+#      MYSQL_ROOT_PASSWORD: password
+#      MYSQL_HOSTNAME : dockerdb
+#      MYSQL_CONTAINER : dockerdb
+ 
+##    expose:
+##      # Opens port 3306 on the container
+##      - '3307'
+
