+++ src/main/resources/application-docker.properties
@@ -1,0 +1,8 @@
+#docker profile
+
+spring.datasource.url=jdbc:mysql://mysql-docker-container:3306/test
+spring.datasource.username=user
+spring.datasource.password =password
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
+server.port=8080
