+++ src/main/resources/application.properties
@@ -1,4 +1,5 @@
-spring.profiles.active=test
+spring.profiles.active=docker
+
